rattler_lock

Module options

Defines SolveOptions and reexports from rattler_solve that are used.

Structs

SolveOptions

Options that were used during the resolution of the packages stored in the lock-file. These options strongly influence the outcome of the solve and are therefore stored along with the locked packages.

Enums

ChannelPriority

Represents the channel priority option to use during solves.

PypiPrereleaseMode

The prerelease mode used to resolve PyPI dependencies.

SolveStrategy

Represents the strategy to use when solving dependencies