rattler_lock

Trait Matches

pub trait Matches<T> {
    // Required method
    fn matches(&self, other: &T) -> bool;
}

A trait that defines the behavior of matching a spec against a record.

Required Methods

fn matches(&self, other: &T) -> bool

Match a [MatchSpec] against a record. Matching it to a record means that the record is valid for the spec.